added users controllers and user form

+++ crm/src/main/java/com/fluidcodes/crm/dao/UsersRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.fluidcodes.crm.entities.Users;
 
-public interface UsersRepo extends JpaRepository<Users,Integer>{
+public interface UsersRepo extends JpaRepository<Users,Long>{
 
 }
